tests

+++ spiral/extension.ps1
@@ -29,23 +29,25 @@ if ($IsWindows -and $env:scoop) {
     }
 }
 
-$spiralPath = "../deps/The-Spiral-Language"
+if ($extensionsPath.Count -gt 0) {
+    $spiralPath = "../deps/The-Spiral-Language"
 
-$extensionSrcPath = "$spiralPath/VS Code Plugin"
+    $extensionSrcPath = "$spiralPath/VS Code Plugin"
 
-$json = Get-Content (Join-Path -Path $extensionSrcPath -ChildPath "package.json") | ConvertFrom-Json
-$vsixName = $json.name + "-" + $json.version + ".vsix"
-$vsixPath = Join-Path -Path $extensionSrcPath -ChildPath $vsixName
+    $json = Get-Content (Join-Path -Path $extensionSrcPath -ChildPath "package.json") | ConvertFrom-Json
+    $vsixName = $json.name + "-" + $json.version + ".vsix"
+    $vsixPath = Join-Path -Path $extensionSrcPath -ChildPath $vsixName
 
-Remove-Item "$extensionSrcPath/compiler" -Recurse -Force -ErrorAction SilentlyContinue
+    Remove-Item "$extensionSrcPath/compiler" -Recurse -Force -ErrorAction SilentlyContinue
 
-Copy-Item -Recurse -Force "$spiralPath/The Spiral Language 2/artifacts/bin/The Spiral Language 2/release/" "$extensionSrcPath/compiler"
+    Copy-Item -Recurse -Force "$spiralPath/The Spiral Language 2/artifacts/bin/The Spiral Language 2/release/" "$extensionSrcPath/compiler"
 
-Set-Location $extensionSrcPath
-npm install
-npx tsc --build
-npx @vscode/vsce package
-Set-Location $ScriptDir
+    Set-Location $extensionSrcPath
+    npm install
+    npx tsc --build
+    npx @vscode/vsce package
+    Set-Location $ScriptDir
+}
 
 $Error | Format-List -Force
 Write-Output "LASTEXITCODE: $LASTEXITCODE"
